BFT and its all operations
What is a BFS tree?
Image result for bft tree
The breadth-first search or BFS algorithm is used to search a tree or graph data structure for a node that meets a set of criteria. It begins at the root of the tree or graph and investigates all nodes at the current depth level before moving on to nodes at the next depth level

In short, we refer to it as BFS traversal. A binary tree is organized into different levels where the root node is at the topmost level (0th level). The idea behind level order traversal is to start by processing the root node, followed by all nodes at the first level, then the second level, and so on.