Deletion of first node in single linked list
A singly linked list is a type of linked list that is unidirectional, that is, it can be traversed in only one direction from head to the last node (tail). Each element in a linked list is called a node. A single node contains data and a pointer to the next node which helps in maintaining the structure of the list.

How to delete a node in singly linked list in C?
Algorithm
STEP 1: IF HEAD = NULL.
STEP 2: SET TEMP = HEAD.
STEP 3: SET I = 0.
STEP 4: REPEAT STEP 5 TO 8 UNTIL I.
STEP 5: TEMP1 = TEMP.
STEP 6: TEMP = TEMP → NEXT.
STEP 7: IF TEMP = NULL.
STEP 8: I = I+1.